requirements.txt
telethon==1.39.0
google-genai==1.13.0
gspread==6.1.2
python-dotenv==1.1.0
PySocks==1.7.1
